Initial Commit, MVP
This project is a front-end for the chat.db file the iMessage
creates. It started because the search functionality in the
Messages app is lacking, if not broken.

Kotlin is used for everything, including the not-yet-1.0
Exposed library for database access.

As a not-yet-1.0 library: documentation is sparse, APIs are likely
to change, and there are bound to be a lot of bugs. But Kotlin
is just so damn sexy, especially the ability library authors
have to make DSL's.

The UI was completely lifted from my Kafka Message Viewer program,
and surprisingly the search functionality is reasonably quick even
when all my messages (~200,000) are loaded in.