Get all messages for a particular chat
I'm able to get all the messages for a particular chat, which
is the first step to grouping the messages in a sane way, but it
is pretty slow and inefficient. I want to stick with using DAOs as
much as possible, but I might need more flexibility in the actual
SQL that gets run.

Right now it's getting all the rows and doing the filtering on the
Kotlin side, but it would be better to have that done as part of
the SQL query.