Refactored DB code to be more performant
Using DAOs seems to be  both inflexible and slow.
All in all, it took several seconds to load a single conversation.
This is wayyyy too slow for a single conversation, nevermind
trying to do multiple.

The apparent inflexibility of DAOs is what causes this slowness.
All of the results are first returned, and then manipulated in
Kotlin, causing unnecessary overhead.

After a lot of messing around, it seems like using the DSL offers
the best balance of speed, safety, and ease of use.

This also allows me to first optimize the SQL queries and then
translate them to the DSL, ensuring the bottleneck is in sqlite3
and the queries themselves rather than in the Kotlin code.