Improved chat bubble appearance and removed caching
List cells can be cached to improve performance and memory
footprint, but I don't know how to do it properly. Trying to use
caching meant that the same few messages were repeated over and
over again.

Removing caching eliminates this problem, but it seems to make
scrolling suck. Instead of being smooth and having inertia, it
is jerky and stops the moment you stop scrolling.

The memory footprint doesn't seem too bad right now. I tested
with 10k messages and it seemed good enough for now, but it
definitely should be addressed soon.