Added attachment and contact info to messages
All of the info is spread across several tables, so most of the
work was figuring out the sequence of joins to get all the info
we want.

The last bit of info we really need - folk's actual names - is
in another database and should come with a future commit.