Clean up
The 'models' folder in the ui package had become a dumping ground.
Moved all the DB tables into the adaptor folder, since that is
the only place they are used. Also moved them all into a single
file since they are so small.

The only "model" we really have is a Message right now, and it's
not UI specific, so it has been refactored to reflect that.