Removed search (for now)
The MessagePane view is getting out of control, and search was
a small part of it. I need to revamp search anyway, so it is the
first casualty in what will be an attempt to wrangle the mess
that the UI code is becoming.

p.s. I think the reason it is a mess is because I still don't
understand how to use a ListCellFragment