Refactored MessagePane and added placeholder image
The unholy mess that was MessagePane has been broken down a bit.
MessageCell was created to hold each individual message, and will
likely need to be broken down further in the future, since
messages can be somewhat complex, and there are several different
types.

For now, display based on properties of the message is handled
with `if` statements and helper functions. Image loading was
put into it's own class and includes some error handling, logging,
and returning a placeholder image in the event of failure.

Failure to load an image that exists implies we need to get
full disk access from the privacy settings under Mojave. I haven't
found any easy way to do this, I think the user has to do it
themselves. I could pop up a dialog or include it in the readme.
Future me will figure it out.