Integrated "Conversations" into the app
A conversation is the sharing of ideas or information between
2 or more people. We want to group messages by conversation.

A model that could make sense is that a conversation is a set
of people and a set of messages. However, within the confines
of this app, it isn't reasonable to store all messages for all
conversations. Therefore, in our model of a conversation, we have
the people with whom we are chatting, and an ID that can be used
to look up the messages in a database. There is also a placeholder
for "last message sent date" to help organize the chats based on
the most recent, but it is unused right now.

Because a conversation revolves around who we are talking to,
the Contact model had to be overhauled with a bunch of helper
methods. We want to see info about who we are chatting with: name
number, etc. So the helper methods are there to accomplish
this goal. It's a bit ugly, but that's because we might only
have an email, or perhaps just a number that isn't even formatted
in a standard way.