Added Conversation Card to add some polish to the UI
Trying to make the UI code look good and do what I want proved
to be too frustrating for the time being. Everything is defined
inline, and there was a lot of fiddling with stuff.

Because this commit is really just concerend with how things
look, I created some placeholder data until we can load real
data.

A by-product of this effort is that now Dates are actually
being parsed now, so that's a plus.