Add lastSentMessage to Conversation
Conversations are loaded in 2 passes to get all the info needed.
Trying to do it in one pass has either been slow, complicated,
or both. But that doesn't mean it's not possible.

The way the last message is being fetched is awfully similar
to the way messages are being fetched (as you might expect).
This similarity makes me thing I ought to be able to re-use more
than I am from the sqliteToMessage file. However, it seems
clear enough and fast enough for now, but definitely worth
revisiting.