Added validation to provide user feedback to DB settings
This provides some validation and user feedback when editing
the db file locations. I think it will address the most common
problems, like permission denied and typos. It won't save the user
from putting in the wrong sqlite database though, as the validation
only checks that it is an accessible sqlite database.