Added logging using log4j2
The most pressing place where logs are useful at the moment is
in the database calls. Having SQL queries with accurate timing
information is helpful for improving those calls.

This PR just lays the foundation for logging, focusing on the
database calls and the log4j2 configuration file. Future PRs can
add logging where necessary now that this foundation is laid.

There is a MacOS specific part of this PR - the log file is placed
in a MacOS default location, which will either not work or just
look weird on other platforms.