Added proof-of-concept button to open photos in googl earth
I am toying with the idea of making a plugin style architecture,
allowing functionality to be added via plugins. This is an attempt
at what I would consider a cool plugin, but it's pretty complex
and brittle. If I expect anyone to ever actually want to use this
app and make plugins for it, I think this is a cool demonstration
of what is possible. Definitely not a good example code-wise though.